Found there was a gap in the SeaFET data from BOB from 7/5/2020 to 7/8/2020 UTC. This happened with a Windows reset over the weekend which halted instrument logging completely once Seabird Inc./Satlantic Software went offline. Perhaps next time I should start logging on a Thursday? Also, I think I should log directly on the instrument and have it transmit to Seabird software in real time...
This was also in part to having computers work after a long coronavirus break.

Regardless, now that I have discovered this I am glad I continued to log for another week. I added a copy of that data to a subdirectory.

I am going to append the original data in my next commit.