Noticed in while preparing to plot the Cal Maritime data that when I assessed the summary of the datetime variable to get a minimum and maximum date that NA's were present for about 1200 variables.
I then went back into the data tidying script and noticed that the data from 10/15 to 12/14 or so were in a MM/DD/YYYY format instead of a YYYY/MM/DD format which threw the datetime variable off.

I went back in and reworked this string to better match the correct format so that when I forced it into POSIXct it created the correct datetime variable.

Now I can begin to properly plot the CMA data